% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{estimate_area}
\alias{estimate_area}
\title{Estimate area function}
\usage{
estimate_area(B = 5000, seed = 10)
}
\arguments{
\item{x}{A non-null positive integer}
}
\value{
Print a list with the area estimation and the points
}
\description{
Estimate area function
}
