% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{plot.area}
\alias{plot.area}
\title{Plot area function}
\usage{
\method{plot}{area}(x)
}
\arguments{
\item{x}{a list provided by estimate_area() function}
}
\value{
plot the estimated area and its points
}
\description{
Plot area function
}
