% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{plot_area}
\alias{plot_area}
\title{Plot area function}
\usage{
plot_area(x)
}
\arguments{
\item{x}{a list provided by estimate_area() function}
}
\value{
Return a plot of the estimated area and its points
}
\description{
Plot area function
}
\examples{
rval <- estimate_area(100)
plot_area(rval)

}
\author{
Corinne & Laurene
}
