% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{estimate_area}
\alias{estimate_area}
\title{Estimate area function}
\usage{
estimate_area(B = 5000, seed = 10)
}
\arguments{
\item{B}{a non-null positive integer}

\item{seed}{an integer or NULL}
}
\value{
Print a list with the area estimation and its points
}
\description{
Estimate area function
}
\examples{
estimate_area(100)

}
\author{
Corinne & Laurene
}
